requirements-dev.txt
fastapi[standard]
pydantic-settings
aiosqlite
itsdangerous
langchain
langchain_community
langchain-core
langchain-text-splitters
langgraph
langgraph-checkpoint-postgres
langgraph-checkpoint-sqlite
langchain-aws
langchain-anthropic
langchain_openai
mcp[cli]
httpx
psycopg[binary]
langchain_mcp_adapters